# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Vendor Compare API for researched, metered product comparisons.

    Customer applications authenticate `/v1` operations with a tenant-scoped API key:
    `Authorization: Bearer vc_live_...`. Account, billing, and API-key management
    operations use a signed-in Clerk session and, where documented, `X-Tenant-Id`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /comparisons:
    get:
      operationId: listComparisons
      tags: [comparisons]
      summary: List the signed-in user's comparison history
      responses:
        "200":
          description: Comparisons from the last 30 days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComparisonSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
    post:
      operationId: createComparison
      tags: [comparisons]
      summary: Create and analyze a new vendor comparison
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonInput"
      responses:
        "201":
          description: Comparison analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "400":
          description: Invalid comparison input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
...
  /comparisons/{id}:
    get:
      operationId: getComparison
      tags: [comparisons]
      summary: Get a comparison analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comparison analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
        "404":
          description: Comparison not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
    delete:
      operationId: deleteComparison
      tags: [comparisons]
      summary: Delete a comparison from history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Comparison deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
        "404":
          description: Comparison not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyErrorResponse"
  /v1/comparisons:
    get:
      operationId: externalListComparisons
      tags: [commercial]
      summary: List tenant comparisons
      description: Requires an API key with the comparisons:read scope.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Tenant comparisons
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Quota-Included: { $ref: "#/components/headers/QuotaIncluded" }
            X-Quota-Used: { $ref: "#/components/headers/QuotaUsed" }
            X-Quota-Remaining: { $ref: "#/components/headers/QuotaRemaining" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ComparisonSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      operationId: externalCreateComparison
      tags: [commercial]
      summary: Create a metered tenant comparison
      description: Requires an API key with the comparisons:write scope and an Idempotency-Key. Only successful completed comparisons consume usage; retries with the same request are replayed.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique key for this request; reusing it with a changed body returns 409.
          schema: { type: string, minLength: 8, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComparisonInput" }
            example:
              prompt: Compare Tesla Model 3 and BYD Seal for an Australian buyer with a A$50,000 budget.
              vendors: [Tesla Model 3, BYD Seal]
              criteria: [Purchase price, Driving range, Warranty, Running costs]
              urls: []
      responses:
        "201":
          description: Completed comparison
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Quota-Included: { $ref: "#/components/headers/QuotaIncluded" }
            X-Quota-Used: { $ref: "#/components/headers/QuotaUsed" }
            X-Quota-Remaining: { $ref: "#/components/headers/QuotaRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comparison" }
              example:
                id: 142
                status: complete
                category: Electric vehicles
                recommendation: BYD Seal
                score: 84
                executiveSummary: BYD Seal provides the stronger budget fit.
                recommendationReason: It remains within budget while preserving range and warranty value.
                vendors: [Tesla Model 3, BYD Seal]
                vendorScores:
                  - { vendor: Tesla Model 3, score: 77, verdict: Strong technology, above budget }
                  - { vendor: BYD Seal, score: 84, verdict: Best overall fit }
                nextSteps: [Confirm drive-away pricing, Book both test drives]
                urls: [https://www.tesla.com/en_au/model3, https://www.bydautomotive.com.au/seal]
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/BadGateway" }
    parameters: []
  /v1/comparisons/{id}:
    get:
      operationId: externalGetComparison
      tags: [commercial]
      summary: Get a tenant comparison
      description: Requires an API key with the comparisons:read scope.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Comparison
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Quota-Included: { $ref: "#/components/headers/QuotaIncluded" }
            X-Quota-Used: { $ref: "#/components/headers/QuotaUsed" }
            X-Quota-Remaining: { $ref: "#/components/headers/QuotaRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comparison" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/usage:
    get:
      operationId: getExternalUsage
      tags: [commercial]
      summary: Get monthly usage and quota
      description: Requires an API key with the usage:read scope.
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: Usage
          headers:
            RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Quota-Included: { $ref: "#/components/headers/QuotaIncluded" }
            X-Quota-Used: { $ref: "#/components/headers/QuotaUsed" }
            X-Quota-Remaining: { $ref: "#/components/headers/QuotaRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Vendor Compare API key
      description: "Tenant-scoped API key. Send Authorization: Bearer vc_live_.... Scope is documented per operation. The plaintext key is returned only when created or rotated."
    ClerkBearer:
      type: http
      scheme: bearer
      bearerFormat: Clerk session
      description: Signed-in Clerk session token for management operations only; do not use this credential for /v1 calls.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  headers:
    RateLimitLimit:
      description: Requests permitted per UTC minute.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in the current UTC minute.
      schema: { type: integer }
    RateLimitReset:
      description: Unix timestamp when the current minute resets.
      schema: { type: integer }
    QuotaIncluded:
      description: Comparisons included in the current provider billing period (or calendar month for inactive/bootstrap display).
      schema: { type: integer }
    QuotaUsed:
      description: Successful comparisons used in the current provider billing period (or calendar month for inactive/bootstrap display).
      schema: { type: integer }
    QuotaRemaining:
      description: Comparisons remaining before the prepaid hard cap is reached.
      schema: { type: integer }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Unauthorized:
      description: Missing or invalid authentication
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Forbidden:
      description: Authentication lacks the required scope or tenant role
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    Conflict:
      description: Idempotency key conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    RateLimited:
      description: Minute rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until retry }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    QuotaExceeded:
      description: Prepaid comparison allowance exhausted
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until the prepaid allowance resets }
        X-Quota-Included: { $ref: "#/components/headers/QuotaIncluded" }
        X-Quota-Used: { $ref: "#/components/headers/QuotaUsed" }
        X-Quota-Remaining: { $ref: "#/components/headers/QuotaRemaining" }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
    BadGateway:
      description: Comparison analysis or durable finalization failed upstream
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
    LegacyErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    ErrorResponse:
      type: object
      description: Stable structured error envelope; error is retained for legacy clients.
      properties:
        code: { type: string, example: invalid_request }
        message: { type: string }
        requestId: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
...
    UsageSummary:
      type: object
      properties:
        month: { type: string, example: "2025-01-15/2025-02-15", description: Provider period label; inactive/bootstrap tenants use YYYY-MM. }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        included: { type: integer }
        used: { type: integer }
        remaining: { type: integer, description: Comparisons remaining before the prepaid hard cap is reached. }
        exhausted: { type: boolean }
      required: [month, periodStart, periodEnd, included, used, remaining, exhausted]
...
    ComparisonInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 8
          maxLength: 2000
        vendors:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: string
            minLength: 1
            maxLength: 120
        urls:
          type: array
          items:
            type: string
            format: uri
        criteria:
          type: array
          maxItems: 8
          items:
            type: string
            minLength: 1
            maxLength: 100
      required: [prompt]
...
    ComparisonSummary:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        vendors:
          type: array
          items:
            type: string
        category:
          type: string
        recommendation:
          type: string
        score:
          type: integer
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [complete, processing, failed]
      required: [id, prompt, vendors, category, recommendation, score, createdAt, status]
    Comparison:
      allOf:
        - $ref: "#/components/schemas/ComparisonSummary"
        - type: object
          properties:
            urls:
              type: array
              items:
                type: string
            criteria:
              type: array
              items:
                type: string
            executiveSummary:
              type: string
            recommendationReason:
              type: string
            vendorScores:
              type: array
              items:
                $ref: "#/components/schemas/VendorScore"
            pricing:
              type: array
              items:
                $ref: "#/components/schemas/AnalysisRow"
            features:
              type: array
              items:
                $ref: "#/components/schemas/AnalysisRow"
            swot:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            opportunities:
              type: array
              items:
                type: string
            insights:
              type: array
              items:
                type: string
            nextSteps:
              type: array
              items:
                type: string
            contextAssumptions:
              type: array
              description: Explicit assumptions made where business, regulatory, security, commercial, operating, integration, data, or maturity context was missing.
              items: { type: string }
            productEquivalency:
              type: array
              description: Like-for-like mapping of current and target products or services, including partial equivalence and uncovered scope.
              items: { $ref: "#/components/schemas/ProductEquivalencyItem" }
            functionalGaps:
              type: array
              description: Required capabilities that are absent, partial, changed, or unverified in the target arrangement.
              items: { $ref: "#/components/schemas/FunctionalGapItem" }
            serviceProductMap:
              type: array
              description: Mapping between business services and the products, dependencies, and owners that enable them.
              items: { $ref: "#/components/schemas/ServiceProductMapItem" }
            migrationSequence:
              type: array
              description: Ordered migration phases with dependencies, exit criteria, and risk.
              items: { $ref: "#/components/schemas/MigrationPhase" }
            decisionGovernance:
              type: array
              description: Decision rights, evidence requirements, approvers, and approval gates.
              items: { $ref: "#/components/schemas/DecisionGovernanceItem" }
          required: [urls, criteria, executiveSummary, recommendationReason, vendorScores, pricing, features, swot, opportunities, insights, nextSteps, contextAssumptions, productEquivalency, functionalGaps, serviceProductMap, migrationSequence, decisionGovernance]
...
    VendorScore:
      type: object
      properties:
        vendor:
          type: string
        score:
          type: integer
        color:
          type: string
        verdict:
          type: string
        weightedScores:
          type: array
          items:
            $ref: "#/components/schemas/WeightedCriterionScore"
        switchConditions:
          type: array
          description: Conditions under which this option should be preferred over the overall recommendation.
          items:
            type: string
        vrio:
          $ref: "#/components/schemas/VrioAssessment"
        marketPosition:
          $ref: "#/components/schemas/MarketPosition"
      required: [vendor, score, color, verdict]
    VrioDimension:
      type: object
      properties:
        status:
          type: string
          enum: [strong, partial, weak, not_applicable]
        rationale:
          type: string
      required: [status, rationale]
    VrioAssessment:
      type: object
      properties:
        value:
          $ref: "#/components/schemas/VrioDimension"
        rarity:
          $ref: "#/components/schemas/VrioDimension"
        imitability:
          $ref: "#/components/schemas/VrioDimension"
        organization:
          $ref: "#/components/schemas/VrioDimension"
        implication:
          type: string
      required: [value, rarity, imitability, organization, implication]
    MarketPosition:
      type: object
      properties:
        marketShare:
          type: string
          description: Latest credible market-share figure or an explicit unavailable statement.
        marketSharePeriod:
          type: string
        market:
          type: string
        shareValue:
          type: string
          description: Public parent-company share price/value when applicable, otherwise Not applicable.
        shareValueAsOf:
          type: string
        applicability:
          type: string
        evidence:
          type: string
      required: [marketShare, marketSharePeriod, market, shareValue, shareValueAsOf, applicability, evidence]
    WeightedCriterionScore:
      type: object
      properties:
        criterion:
          type: string
        weight:
          type: integer
        score:
          type: integer
          minimum: 0
          maximum: 100
        rationale:
          type: string
      required: [criterion, weight, score, rationale]
    AnalysisRow:
      type: object
      properties:
        dimension:
          type: string
        values:
          type: object
          additionalProperties:
            type: string
        winner:
          type: string
      required: [dimension, values, winner]
    ProductEquivalencyItem:
      type: object
      properties:
        capability: { type: string }
        currentArrangement: { type: string }
        targetArrangement: { type: string }
        equivalency: { type: string, description: Full, partial, no direct equivalent, or unverified, with rationale. }
        gap: { type: string }
      required: [capability, currentArrangement, targetArrangement, equivalency, gap]
    FunctionalGapItem:
      type: object
      properties:
        capability: { type: string }
        currentState: { type: string }
        targetState: { type: string }
        gap: { type: string }
        mitigation: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
      required: [capability, currentState, targetState, gap, mitigation, severity]
    ServiceProductMapItem:
      type: object
      properties:
        businessService: { type: string }
        currentProduct: { type: string }
        targetProduct: { type: string }
        dependencies: { type: string }
        owner: { type: string }
      required: [businessService, currentProduct, targetProduct, dependencies, owner]
    MigrationPhase:
      type: object
      properties:
        phase: { type: string }
        objective: { type: string }
        dependencies: { type: string }
        exitCriteria: { type: string }
        risk: { type: string, enum: [low, medium, high, critical] }
      required: [phase, objective, dependencies, exitCriteria, risk]
    DecisionGovernanceItem:
      type: object
      properties:
        decision: { type: string }
        owner: { type: string }
        approvers: { type: string }
        evidenceRequired: { type: string }
        decisionGate: { type: string }
      required: [decision, owner, approvers, evidenceRequired, decisionGate]
...