# lib/api-zod/src/generated/types/comparisonInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Vendor Compare API for researched, metered product comparisons.
 *
 * Customer applications authenticate `/v1` operations with a tenant-scoped API key:
 * `Authorization: Bearer vc_live_...`. Account, billing, and API-key management
 * operations use a signed-in Clerk session and, where documented, `X-Tenant-Id`.
 *
 * OpenAPI spec version: 0.1.0
 */

export interface ComparisonInput {
  /**
     * @minLength 8
     * @maxLength 2000
     */
  prompt: string;
  /**
     * @minItems 2
     * @maxItems 5
     * @items.minLength 1
     * @items.maxLength 120
     */
  vendors?: string[];
  urls?: string[];
  /**
     * @maxItems 8
     * @items.minLength 1
     * @items.maxLength 100
     */
  criteria?: string[];
}