# lib/api-zod/src/generated/types/vrioDimension.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Vendor Compare API for researched, metered product comparisons.
 *
 * Customer applications authenticate `/v1` operations with a tenant-scoped API key:
 * `Authorization: Bearer vc_live_...`. Account, billing, and API-key management
 * operations use a signed-in Clerk session and, where documented, `X-Tenant-Id`.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { VrioDimensionStatus } from './vrioDimensionStatus';

export interface VrioDimension {
  status: VrioDimensionStatus;
  rationale: string;
}